python/sandbox/pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergey Royz",email = "zjor.se@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.2.2,<3.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
